Crate rcudnn

Provides a safe and convenient wrapper for the CUDA cuDNN API.

This crate (1.0.0) was developed against cuDNN v3.

Architecture

This crate provides three levels of entrace.

FFI
The ffi module exposes the foreign function interface and cuDNN specific types. Usually, there should be no use to touch it if you only want to use cuDNN in you application. The ffi is provided by the rust-cudnn-sys crate and gets reexported here.

Low-Level
The api module exposes already a complete and safe wrapper for the cuDNN API, including proper Rust Errors. Usually there should be not need to use the API directly though, as the Cudnn module, as described in the next block, provides all the API functionality but provides a more convenient interface.

High-Level
The cudnn module exposes the Cudnn struct, which provides a very convenient, easy-to-understand interface for the cuDNN API. There should be not much need to obtain and read the cuDNN manual. Initialize the Cudnn struct and you can call the available methods wich are representing all the available cuDNN operations.

Examples

extern crate rcudnn as cudnn;
extern crate libc;
use cudnn::{Cudnn, TensorDescriptor};
use cudnn::utils::{ScalParams, DataType};
fn main() {
//  Initialize a new cuDNN context and allocates resources.
    let cudnn = Cudnn::new().unwrap();
//  Create a cuDNN Tensor Descriptor for `src` and `dest` memory.
    let src_desc = TensorDescriptor::new(&[2, 2, 2], &[4, 2, 1], DataType::Float).unwrap();
    let dest_desc = TensorDescriptor::new(&[2, 2, 2], &[4, 2, 1], DataType::Float).unwrap();
    let acti = cudnn.init_activation().unwrap();
//  Obtain the `src` and memory pointer on the GPU.
//  NOTE: You wouldn't do it like that. You need to really allocate memory on the GPU with e.g. CUDA or Collenchyma.
    let src_data: *const ::libc::c_void = ::std::ptr::null();
    let dest_data: *mut ::libc::c_void = ::std::ptr::null_mut();
//  Now you can compute the forward sigmoid activation on your GPU.
    cudnn.sigmoid_forward::<f32>(&acti, &src_desc, src_data, &dest_desc, dest_data, ScalParams::default());
}

Notes

rust-cudnn was developed at Autumn for the Rust Machine Intelligence Framework Leaf.

rust-cudnn is part of the High-Performance Computation Framework Collenchyma, for the Neural Network Plugin. Rust CUDNN is now maintained by Juice

Modules

cuda

Defines Cuda Device Memory.

utils

Describes utility functionality for CUDA cuDNN.

Structs

API

Defines the Cuda cuDNN API.

ActivationDescriptor

Describes a ActivationDescriptor.

CUevent_st

CUexternalMemory_st

CUexternalSemaphore_st

CUfunc_st

CUgraphExec_st

CUgraphNode_st

CUgraph_st

CUmemPoolHandle_st

CUstream_st

CUuserObject_st

CUuuid_st

ConvolutionDescriptor

Describes a Convolution Descriptor.

Cudnn

Provides a the high-level interface to CUDA’s cuDNN.

DropoutDescriptor

Describes a DropoutDescriptor.

FilterDescriptor

Describes a Filter Descriptor.

NormalizationDescriptor

Describes a LRN Descriptor.

PoolingDescriptor

Describes a Pooling Descriptor.

RnnDescriptor

Describes a Recurrent Descriptor.

TensorDescriptor

Describes a TensorDescriptor.

cudaAccessPolicyWindow

cudaArray

cudaArrayMemoryRequirements

cudaArraySparseProperties

cudaArraySparseProperties__bindgen_ty_1

cudaChannelFormatDesc

cudaDeviceProp

cudaExtent

cudaExternalMemoryBufferDesc

cudaExternalMemoryHandleDesc

cudaExternalMemoryHandleDesc__bindgen_ty_1__bindgen_ty_1

cudaExternalMemoryMipmappedArrayDesc

cudaExternalSemaphoreHandleDesc

cudaExternalSemaphoreHandleDesc__bindgen_ty_1__bindgen_ty_1

cudaExternalSemaphoreSignalNodeParams

cudaExternalSemaphoreSignalParams

cudaExternalSemaphoreSignalParams__bindgen_ty_1

cudaExternalSemaphoreSignalParams__bindgen_ty_1__bindgen_ty_1

cudaExternalSemaphoreSignalParams__bindgen_ty_1__bindgen_ty_3

cudaExternalSemaphoreSignalParams_v1

cudaExternalSemaphoreSignalParams_v1__bindgen_ty_1

cudaExternalSemaphoreSignalParams_v1__bindgen_ty_1__bindgen_ty_1

cudaExternalSemaphoreSignalParams_v1__bindgen_ty_1__bindgen_ty_3

cudaExternalSemaphoreWaitNodeParams

cudaExternalSemaphoreWaitParams

cudaExternalSemaphoreWaitParams__bindgen_ty_1

cudaExternalSemaphoreWaitParams__bindgen_ty_1__bindgen_ty_1

cudaExternalSemaphoreWaitParams__bindgen_ty_1__bindgen_ty_3

cudaExternalSemaphoreWaitParams_v1

cudaExternalSemaphoreWaitParams_v1__bindgen_ty_1

cudaExternalSemaphoreWaitParams_v1__bindgen_ty_1__bindgen_ty_1

cudaExternalSemaphoreWaitParams_v1__bindgen_ty_1__bindgen_ty_3

cudaFuncAttributes

cudaGraphicsResource

cudaHostNodeParams

cudaIpcEventHandle_st

cudaIpcMemHandle_st

cudaKernelNodeParams

cudaLaunchParams

cudaMemAccessDesc

cudaMemAllocNodeParams

cudaMemLocation

cudaMemPoolProps

cudaMemPoolPtrExportData

cudaMemcpy3DParms

cudaMemcpy3DPeerParms

cudaMemsetParams

cudaMipmappedArray

cudaPitchedPtr

cudaPointerAttributes

cudaPos

cudaResourceDesc

cudaResourceDesc__bindgen_ty_1__bindgen_ty_1

cudaResourceDesc__bindgen_ty_1__bindgen_ty_2

cudaResourceDesc__bindgen_ty_1__bindgen_ty_3

cudaResourceDesc__bindgen_ty_1__bindgen_ty_4

cudaResourceViewDesc

cudaTextureDesc

cudnnActivationStruct

cudnnAlgorithmPerformanceStruct

cudnnAlgorithmStruct

cudnnAlgorithmUnionStruct

cudnnAttnStruct

cudnnCTCLossStruct

cudnnContext

cudnnConvolutionBwdDataAlgoPerfStruct

cudnnConvolutionBwdFilterAlgoPerfStruct

cudnnConvolutionFwdAlgoPerfStruct

cudnnConvolutionStruct

cudnnDebugStruct

cudnnDropoutStruct

cudnnFilterStruct

cudnnFusedOpsConstParamStruct

cudnnFusedOpsPlanStruct

cudnnFusedOpsVariantParamStruct

cudnnLRNStruct

cudnnOpTensorStruct

cudnnPersistentRNNPlan

cudnnPoolingStruct

cudnnRNNDataStruct

cudnnRNNStruct

cudnnReduceTensorStruct

cudnnRuntimeTag_t

cudnnSeqDataStruct

cudnnSpatialTransformerStruct

cudnnTensorStruct

cudnnTensorTransformStruct

dim3

surfaceReference

textureReference

Enums

Error

Defines CUDA’s cuDNN errors.

cudaAccessProperty

cudaCGScope

cudaChannelFormatKind

cudaComputeMode

cudaDataType

cudaDataType_t

cudaDeviceAttr

cudaDeviceP2PAttr

cudaError

cudaError_t

cudaExternalMemoryHandleType

cudaExternalSemaphoreHandleType

cudaFlushGPUDirectRDMAWritesOptions

cudaFlushGPUDirectRDMAWritesScope

cudaFlushGPUDirectRDMAWritesTarget

cudaFuncAttribute

cudaFuncCache

cudaGPUDirectRDMAWritesOrdering

cudaGetDriverEntryPointFlags

cudaGraphDebugDotFlags

cudaGraphExecUpdateResult

cudaGraphInstantiateFlags

cudaGraphMemAttributeType

cudaGraphNodeType

cudaGraphicsCubeFace

cudaGraphicsMapFlags

cudaGraphicsRegisterFlags

cudaKernelNodeAttrID

cudaLimit

cudaMemAccessFlags

cudaMemAllocationHandleType

cudaMemAllocationType

cudaMemLocationType

cudaMemPoolAttr

cudaMemRangeAttribute

cudaMemcpyKind

cudaMemoryAdvise

cudaMemoryType

cudaOutputMode

cudaOutputMode_t

cudaResourceType

cudaResourceViewFormat

cudaRoundMode

cudaSharedCarveout

cudaSharedMemConfig

cudaStreamAttrID

cudaStreamCaptureMode

cudaStreamCaptureStatus

cudaStreamUpdateCaptureDependenciesFlags

cudaSurfaceBoundaryMode

cudaSurfaceFormatMode

cudaSynchronizationPolicy

cudaTextureAddressMode

cudaTextureFilterMode

cudaTextureReadMode

cudaUserObjectFlags

cudaUserObjectRetainFlags

cudnnActivationMode_t

cudnnBackendAttributeName_t

cudnnBackendAttributeType_t

cudnnBackendBehaviorNote_t

cudnnBackendDescriptorType_t

cudnnBackendHeurMode_t

cudnnBackendKnobType_t

cudnnBackendLayoutType_t

cudnnBackendNumericalNote_t

cudnnBackendTensorReordering_t

cudnnBatchNormMode_t

cudnnBatchNormOps_t

cudnnBnFinalizeStatsMode_t

cudnnCTCLossAlgo_t

cudnnConvolutionBwdDataAlgo_t

cudnnConvolutionBwdFilterAlgo_t

cudnnConvolutionFwdAlgo_t

cudnnConvolutionMode_t

cudnnDataType_t

cudnnDeterminism_t

cudnnDirectionMode_t

cudnnDivNormMode_t

cudnnErrQueryMode_t

cudnnFoldingDirection_t

cudnnForwardMode_t

cudnnFusedOpsConstParamLabel_t

cudnnFusedOpsPointerPlaceHolder_t

cudnnFusedOpsVariantParamLabel_t

cudnnFusedOps_t

cudnnGenStatsMode_t

cudnnIndicesType_t

cudnnLRNMode_t

cudnnLossNormalizationMode_t

cudnnMathType_t

cudnnMultiHeadAttnWeightKind_t

cudnnNanPropagation_t

cudnnNormAlgo_t

cudnnNormMode_t

cudnnNormOps_t

cudnnOpTensorOp_t

cudnnPaddingMode_t

cudnnPointwiseMode_t

cudnnPoolingMode_t

cudnnRNNAlgo_t

cudnnRNNBiasMode_t

cudnnRNNClipMode_t

cudnnRNNDataLayout_t

cudnnRNNInputMode_t

cudnnRNNMode_t

cudnnReduceTensorIndices_t

cudnnReduceTensorOp_t

cudnnReorderType_t

cudnnResampleMode_t

cudnnSamplerType_t

cudnnSeqDataAxis_t

cudnnSeverity_t

cudnnSoftmaxAlgorithm_t

cudnnSoftmaxMode_t

cudnnStatus_t

cudnnTensorFormat_t

cudnnWgradMode_t

Constants

CUDNN_ADV_INFER_MAJOR

CUDNN_ADV_INFER_MINOR

CUDNN_ADV_INFER_PATCH

CUDNN_ADV_TRAIN_MAJOR

CUDNN_ADV_TRAIN_MINOR

CUDNN_ADV_TRAIN_PATCH

CUDNN_ATTN_DISABLE_PROJ_BIASES

CUDNN_ATTN_ENABLE_PROJ_BIASES

CUDNN_ATTN_QUERYMAP_ALL_TO_ONE

CUDNN_ATTN_QUERYMAP_ONE_TO_ONE

CUDNN_ATTN_WKIND_COUNT

CUDNN_BN_MIN_EPSILON

CUDNN_CNN_INFER_MAJOR

CUDNN_CNN_INFER_MINOR

CUDNN_CNN_INFER_PATCH

CUDNN_CNN_TRAIN_MAJOR

CUDNN_CNN_TRAIN_MINOR

CUDNN_CNN_TRAIN_PATCH

CUDNN_DIM_MAX

CUDNN_LRN_MAX_N

CUDNN_LRN_MIN_BETA

CUDNN_LRN_MIN_K

CUDNN_LRN_MIN_N

CUDNN_MAJOR

CUDNN_MINOR

CUDNN_OPS_INFER_MAJOR

CUDNN_OPS_INFER_MINOR

CUDNN_OPS_INFER_PATCH

CUDNN_OPS_TRAIN_MAJOR

CUDNN_OPS_TRAIN_MINOR

CUDNN_OPS_TRAIN_PATCH

CUDNN_PATCHLEVEL

CUDNN_RNN_PADDED_IO_DISABLED

CUDNN_RNN_PADDED_IO_ENABLED

CUDNN_SEQDATA_DIM_COUNT

CUDNN_VERSION

libraryPropertyType_t_MAJOR_VERSION

libraryPropertyType_t_MINOR_VERSION

libraryPropertyType_t_PATCH_LEVEL

Functions

cudaArrayGetInfo⚠

cudaArrayGetMemoryRequirements⚠

cudaArrayGetPlane⚠

cudaArrayGetSparseProperties⚠

cudaBindSurfaceToArray⚠

cudaBindTexture⚠

cudaBindTexture2D⚠

cudaBindTextureToArray⚠

cudaBindTextureToMipmappedArray⚠

cudaChooseDevice⚠

cudaCreateChannelDesc⚠

cudaCreateSurfaceObject⚠

cudaCreateTextureObject⚠

cudaCtxResetPersistingL2Cache⚠

cudaDestroyExternalMemory⚠

cudaDestroyExternalSemaphore⚠

cudaDestroySurfaceObject⚠

cudaDestroyTextureObject⚠

cudaDeviceCanAccessPeer⚠

cudaDeviceDisablePeerAccess⚠

cudaDeviceEnablePeerAccess⚠

cudaDeviceFlushGPUDirectRDMAWrites⚠

cudaDeviceGetAttribute⚠

cudaDeviceGetByPCIBusId⚠

cudaDeviceGetCacheConfig⚠

cudaDeviceGetDefaultMemPool⚠

cudaDeviceGetGraphMemAttribute⚠

cudaDeviceGetLimit⚠

cudaDeviceGetMemPool⚠

cudaDeviceGetNvSciSyncAttributes⚠

cudaDeviceGetP2PAttribute⚠

cudaDeviceGetPCIBusId⚠

cudaDeviceGetSharedMemConfig⚠

cudaDeviceGetStreamPriorityRange⚠

cudaDeviceGetTexture1DLinearMaxWidth⚠

cudaDeviceGraphMemTrim⚠

cudaDeviceReset⚠

cudaDeviceSetCacheConfig⚠

cudaDeviceSetGraphMemAttribute⚠

cudaDeviceSetLimit⚠

cudaDeviceSetMemPool⚠

cudaDeviceSetSharedMemConfig⚠

cudaDeviceSynchronize⚠

cudaDriverGetVersion⚠

cudaEventCreate⚠

cudaEventCreateWithFlags⚠

cudaEventDestroy⚠

cudaEventElapsedTime⚠

cudaEventQuery⚠

cudaEventRecord⚠

cudaEventRecordWithFlags⚠

cudaEventSynchronize⚠

cudaExternalMemoryGetMappedBuffer⚠

cudaExternalMemoryGetMappedMipmappedArray⚠

cudaFree⚠

cudaFreeArray⚠

cudaFreeAsync⚠

cudaFreeHost⚠

cudaFreeMipmappedArray⚠

cudaFuncGetAttributes⚠

cudaFuncSetAttribute⚠

cudaFuncSetCacheConfig⚠

cudaFuncSetSharedMemConfig⚠

cudaGetChannelDesc⚠

cudaGetDevice⚠

cudaGetDeviceCount⚠

cudaGetDeviceFlags⚠

cudaGetDeviceProperties⚠

cudaGetDriverEntryPoint⚠

cudaGetErrorName⚠

cudaGetErrorString⚠

cudaGetExportTable⚠

cudaGetFuncBySymbol⚠

cudaGetLastError⚠

cudaGetMipmappedArrayLevel⚠

cudaGetSurfaceObjectResourceDesc⚠

cudaGetSurfaceReference⚠

cudaGetSymbolAddress⚠

cudaGetSymbolSize⚠

cudaGetTextureAlignmentOffset⚠

cudaGetTextureObjectResourceDesc⚠

cudaGetTextureObjectResourceViewDesc⚠

cudaGetTextureObjectTextureDesc⚠

cudaGetTextureReference⚠

cudaGraphAddChildGraphNode⚠

cudaGraphAddDependencies⚠

cudaGraphAddEmptyNode⚠

cudaGraphAddEventRecordNode⚠

cudaGraphAddEventWaitNode⚠

cudaGraphAddExternalSemaphoresSignalNode⚠

cudaGraphAddExternalSemaphoresWaitNode⚠

cudaGraphAddHostNode⚠

cudaGraphAddKernelNode⚠

cudaGraphAddMemAllocNode⚠

cudaGraphAddMemFreeNode⚠

cudaGraphAddMemcpyNode⚠

cudaGraphAddMemcpyNode1D⚠

cudaGraphAddMemcpyNodeFromSymbol⚠

cudaGraphAddMemcpyNodeToSymbol⚠

cudaGraphAddMemsetNode⚠

cudaGraphChildGraphNodeGetGraph⚠

cudaGraphClone⚠

cudaGraphCreate⚠

cudaGraphDebugDotPrint⚠

cudaGraphDestroy⚠

cudaGraphDestroyNode⚠

cudaGraphEventRecordNodeGetEvent⚠

cudaGraphEventRecordNodeSetEvent⚠

cudaGraphEventWaitNodeGetEvent⚠

cudaGraphEventWaitNodeSetEvent⚠

cudaGraphExecChildGraphNodeSetParams⚠

cudaGraphExecDestroy⚠

cudaGraphExecEventRecordNodeSetEvent⚠

cudaGraphExecEventWaitNodeSetEvent⚠

cudaGraphExecExternalSemaphoresSignalNodeSetParams⚠

cudaGraphExecExternalSemaphoresWaitNodeSetParams⚠

cudaGraphExecHostNodeSetParams⚠

cudaGraphExecKernelNodeSetParams⚠

cudaGraphExecMemcpyNodeSetParams⚠

cudaGraphExecMemcpyNodeSetParams1D⚠

cudaGraphExecMemcpyNodeSetParamsFromSymbol⚠

cudaGraphExecMemcpyNodeSetParamsToSymbol⚠

cudaGraphExecMemsetNodeSetParams⚠

cudaGraphExecUpdate⚠

cudaGraphExternalSemaphoresSignalNodeGetParams⚠

cudaGraphExternalSemaphoresSignalNodeSetParams⚠

cudaGraphExternalSemaphoresWaitNodeGetParams⚠

cudaGraphExternalSemaphoresWaitNodeSetParams⚠

cudaGraphGetEdges⚠

cudaGraphGetNodes⚠

cudaGraphGetRootNodes⚠

cudaGraphHostNodeGetParams⚠

cudaGraphHostNodeSetParams⚠

cudaGraphInstantiate⚠

cudaGraphInstantiateWithFlags⚠

cudaGraphKernelNodeCopyAttributes⚠

cudaGraphKernelNodeGetAttribute⚠

cudaGraphKernelNodeGetParams⚠

cudaGraphKernelNodeSetAttribute⚠

cudaGraphKernelNodeSetParams⚠

cudaGraphLaunch⚠

cudaGraphMemAllocNodeGetParams⚠

cudaGraphMemFreeNodeGetParams⚠

cudaGraphMemcpyNodeGetParams⚠

cudaGraphMemcpyNodeSetParams⚠

cudaGraphMemcpyNodeSetParams1D⚠

cudaGraphMemcpyNodeSetParamsFromSymbol⚠

cudaGraphMemcpyNodeSetParamsToSymbol⚠

cudaGraphMemsetNodeGetParams⚠

cudaGraphMemsetNodeSetParams⚠

cudaGraphNodeFindInClone⚠

cudaGraphNodeGetDependencies⚠

cudaGraphNodeGetDependentNodes⚠

cudaGraphNodeGetEnabled⚠

cudaGraphNodeGetType⚠

cudaGraphNodeSetEnabled⚠

cudaGraphReleaseUserObject⚠

cudaGraphRemoveDependencies⚠

cudaGraphRetainUserObject⚠

cudaGraphUpload⚠

cudaGraphicsMapResources⚠

cudaGraphicsResourceGetMappedMipmappedArray⚠

cudaGraphicsResourceGetMappedPointer⚠

cudaGraphicsResourceSetMapFlags⚠

cudaGraphicsSubResourceGetMappedArray⚠

cudaGraphicsUnmapResources⚠

cudaGraphicsUnregisterResource⚠

cudaHostAlloc⚠

cudaHostGetDevicePointer⚠

cudaHostGetFlags⚠

cudaHostRegister⚠

cudaHostUnregister⚠

cudaImportExternalMemory⚠

cudaImportExternalSemaphore⚠

cudaIpcCloseMemHandle⚠

cudaIpcGetEventHandle⚠

cudaIpcGetMemHandle⚠

cudaIpcOpenEventHandle⚠

cudaIpcOpenMemHandle⚠

cudaLaunchCooperativeKernel⚠

cudaLaunchCooperativeKernelMultiDevice⚠

cudaLaunchHostFunc⚠

cudaLaunchKernel⚠

cudaMalloc⚠

cudaMalloc3D⚠

cudaMalloc3DArray⚠

cudaMallocArray⚠

cudaMallocAsync⚠

cudaMallocFromPoolAsync⚠

cudaMallocHost⚠

cudaMallocManaged⚠

cudaMallocMipmappedArray⚠

cudaMallocPitch⚠

cudaMemAdvise⚠

cudaMemGetInfo⚠

cudaMemPoolCreate⚠

cudaMemPoolDestroy⚠

cudaMemPoolExportPointer⚠

cudaMemPoolExportToShareableHandle⚠

cudaMemPoolGetAccess⚠

cudaMemPoolGetAttribute⚠

cudaMemPoolImportFromShareableHandle⚠

cudaMemPoolImportPointer⚠

cudaMemPoolSetAccess⚠

cudaMemPoolSetAttribute⚠

cudaMemPoolTrimTo⚠

cudaMemPrefetchAsync⚠

cudaMemRangeGetAttribute⚠

cudaMemRangeGetAttributes⚠

cudaMemcpy⚠

cudaMemcpy2D⚠

cudaMemcpy2DArrayToArray⚠

cudaMemcpy2DAsync⚠

cudaMemcpy2DFromArray⚠

cudaMemcpy2DFromArrayAsync⚠

cudaMemcpy2DToArray⚠

cudaMemcpy2DToArrayAsync⚠

cudaMemcpy3D⚠

cudaMemcpy3DAsync⚠

cudaMemcpy3DPeer⚠

cudaMemcpy3DPeerAsync⚠

cudaMemcpyArrayToArray⚠

cudaMemcpyAsync⚠

cudaMemcpyFromArray⚠

cudaMemcpyFromArrayAsync⚠

cudaMemcpyFromSymbol⚠

cudaMemcpyFromSymbolAsync⚠

cudaMemcpyPeer⚠

cudaMemcpyPeerAsync⚠

cudaMemcpyToArray⚠

cudaMemcpyToArrayAsync⚠

cudaMemcpyToSymbol⚠

cudaMemcpyToSymbolAsync⚠

cudaMemset⚠

cudaMemset2D⚠

cudaMemset2DAsync⚠

cudaMemset3D⚠

cudaMemset3DAsync⚠

cudaMemsetAsync⚠

cudaMipmappedArrayGetMemoryRequirements⚠

cudaMipmappedArrayGetSparseProperties⚠

cudaOccupancyAvailableDynamicSMemPerBlock⚠

cudaOccupancyMaxActiveBlocksPerMultiprocessor⚠

cudaOccupancyMaxActiveBlocksPerMultiprocessorWithFlags⚠

cudaPeekAtLastError⚠

cudaPointerGetAttributes⚠

cudaRuntimeGetVersion⚠

cudaSetDevice⚠

cudaSetDeviceFlags⚠

cudaSetDoubleForDevice⚠

cudaSetDoubleForHost⚠

cudaSetValidDevices⚠

cudaSignalExternalSemaphoresAsync_v2⚠

cudaStreamAddCallback⚠

cudaStreamAttachMemAsync⚠

cudaStreamBeginCapture⚠

cudaStreamCopyAttributes⚠

cudaStreamCreate⚠

cudaStreamCreateWithFlags⚠

cudaStreamCreateWithPriority⚠

cudaStreamDestroy⚠

cudaStreamEndCapture⚠

cudaStreamGetAttribute⚠

cudaStreamGetCaptureInfo⚠

cudaStreamGetCaptureInfo_v2⚠

cudaStreamGetFlags⚠

cudaStreamGetPriority⚠

cudaStreamIsCapturing⚠

cudaStreamQuery⚠

cudaStreamSetAttribute⚠

cudaStreamSynchronize⚠

cudaStreamUpdateCaptureDependencies⚠

cudaStreamWaitEvent⚠

cudaThreadExchangeStreamCaptureMode⚠

cudaThreadExit⚠

cudaThreadGetCacheConfig⚠

cudaThreadGetLimit⚠

cudaThreadSetCacheConfig⚠

cudaThreadSetLimit⚠

cudaThreadSynchronize⚠

cudaUnbindTexture⚠

cudaUserObjectCreate⚠

cudaUserObjectRelease⚠

cudaUserObjectRetain⚠

cudaWaitExternalSemaphoresAsync_v2⚠

cudnnActivationBackward⚠

cudnnActivationForward⚠

cudnnAddTensor⚠

cudnnAdvInferVersionCheck⚠

cudnnAdvTrainVersionCheck⚠

cudnnBackendCreateDescriptor⚠

cudnnBackendDestroyDescriptor⚠

cudnnBackendExecute⚠

cudnnBackendFinalize⚠

cudnnBackendGetAttribute⚠

cudnnBackendInitialize⚠

cudnnBackendSetAttribute⚠

cudnnBatchNormalizationBackward⚠

cudnnBatchNormalizationBackwardEx⚠

cudnnBatchNormalizationForwardInference⚠

cudnnBatchNormalizationForwardTraining⚠

cudnnBatchNormalizationForwardTrainingEx⚠

cudnnBuildRNNDynamic⚠

cudnnCTCLoss⚠

cudnnCTCLoss_v8⚠

cudnnCnnInferVersionCheck⚠

cudnnCnnTrainVersionCheck⚠

cudnnConvolutionBackwardBias⚠

cudnnConvolutionBackwardData⚠

cudnnConvolutionBackwardFilter⚠

cudnnConvolutionBiasActivationForward⚠

cudnnConvolutionForward⚠

cudnnCopyAlgorithmDescriptor⚠

cudnnCreate⚠

cudnnCreateActivationDescriptor⚠

cudnnCreateAlgorithmDescriptor⚠

cudnnCreateAlgorithmPerformance⚠

cudnnCreateAttnDescriptor⚠

cudnnCreateCTCLossDescriptor⚠

cudnnCreateConvolutionDescriptor⚠

cudnnCreateDropoutDescriptor⚠

cudnnCreateFilterDescriptor⚠

cudnnCreateFusedOpsConstParamPack⚠

cudnnCreateFusedOpsPlan⚠

cudnnCreateFusedOpsVariantParamPack⚠

cudnnCreateLRNDescriptor⚠

cudnnCreateOpTensorDescriptor⚠

cudnnCreatePersistentRNNPlan⚠

cudnnCreatePoolingDescriptor⚠

cudnnCreateRNNDataDescriptor⚠

cudnnCreateRNNDescriptor⚠

cudnnCreateReduceTensorDescriptor⚠

cudnnCreateSeqDataDescriptor⚠

cudnnCreateSpatialTransformerDescriptor⚠

cudnnCreateTensorDescriptor⚠

cudnnCreateTensorTransformDescriptor⚠

cudnnDeriveBNTensorDescriptor⚠

cudnnDeriveNormTensorDescriptor⚠

cudnnDestroy⚠

cudnnDestroyActivationDescriptor⚠

cudnnDestroyAlgorithmDescriptor⚠

cudnnDestroyAlgorithmPerformance⚠

cudnnDestroyAttnDescriptor⚠

cudnnDestroyCTCLossDescriptor⚠

cudnnDestroyConvolutionDescriptor⚠

cudnnDestroyDropoutDescriptor⚠

cudnnDestroyFilterDescriptor⚠

cudnnDestroyFusedOpsConstParamPack⚠

cudnnDestroyFusedOpsPlan⚠

cudnnDestroyFusedOpsVariantParamPack⚠

cudnnDestroyLRNDescriptor⚠

cudnnDestroyOpTensorDescriptor⚠

cudnnDestroyPersistentRNNPlan⚠

cudnnDestroyPoolingDescriptor⚠

cudnnDestroyRNNDataDescriptor⚠

cudnnDestroyRNNDescriptor⚠

cudnnDestroyReduceTensorDescriptor⚠

cudnnDestroySeqDataDescriptor⚠

cudnnDestroySpatialTransformerDescriptor⚠

cudnnDestroyTensorDescriptor⚠

cudnnDestroyTensorTransformDescriptor⚠

cudnnDivisiveNormalizationBackward⚠

cudnnDivisiveNormalizationForward⚠

cudnnDropoutBackward⚠

cudnnDropoutForward⚠

cudnnDropoutGetReserveSpaceSize⚠

cudnnDropoutGetStatesSize⚠

cudnnFindConvolutionBackwardDataAlgorithm⚠

cudnnFindConvolutionBackwardDataAlgorithmEx⚠

cudnnFindConvolutionBackwardFilterAlgorithm⚠

cudnnFindConvolutionBackwardFilterAlgorithmEx⚠

cudnnFindConvolutionForwardAlgorithm⚠

cudnnFindConvolutionForwardAlgorithmEx⚠

cudnnFindRNNBackwardDataAlgorithmEx⚠

cudnnFindRNNBackwardWeightsAlgorithmEx⚠

cudnnFindRNNForwardInferenceAlgorithmEx⚠

cudnnFindRNNForwardTrainingAlgorithmEx⚠

cudnnFusedOpsExecute⚠

cudnnGetActivationDescriptor⚠

cudnnGetActivationDescriptorSwishBeta⚠

cudnnGetAlgorithmDescriptor⚠

cudnnGetAlgorithmPerformance⚠

cudnnGetAlgorithmSpaceSize⚠

cudnnGetAttnDescriptor⚠

cudnnGetBatchNormalizationBackwardExWorkspaceSize⚠

cudnnGetBatchNormalizationForwardTrainingExWorkspaceSize⚠

cudnnGetBatchNormalizationTrainingExReserveSpaceSize⚠

cudnnGetCTCLossDescriptor⚠

cudnnGetCTCLossDescriptorEx⚠

cudnnGetCTCLossDescriptor_v8⚠

cudnnGetCTCLossWorkspaceSize⚠

cudnnGetCTCLossWorkspaceSize_v8⚠

cudnnGetCallback⚠

cudnnGetConvolution2dDescriptor⚠

cudnnGetConvolution2dForwardOutputDim⚠

cudnnGetConvolutionBackwardDataAlgorithmMaxCount⚠

cudnnGetConvolutionBackwardDataAlgorithm_v7⚠

cudnnGetConvolutionBackwardDataWorkspaceSize⚠

cudnnGetConvolutionBackwardFilterAlgorithmMaxCount⚠

cudnnGetConvolutionBackwardFilterAlgorithm_v7⚠

cudnnGetConvolutionBackwardFilterWorkspaceSize⚠

cudnnGetConvolutionForwardAlgorithmMaxCount⚠

cudnnGetConvolutionForwardAlgorithm_v7⚠

cudnnGetConvolutionForwardWorkspaceSize⚠

cudnnGetConvolutionGroupCount⚠

cudnnGetConvolutionMathType⚠

cudnnGetConvolutionNdDescriptor⚠

cudnnGetConvolutionNdForwardOutputDim⚠

cudnnGetConvolutionReorderType⚠

cudnnGetCudartVersion⚠

cudnnGetDropoutDescriptor⚠

cudnnGetErrorString⚠

cudnnGetFilter4dDescriptor⚠

cudnnGetFilterNdDescriptor⚠

cudnnGetFilterSizeInBytes⚠

cudnnGetFoldedConvBackwardDataDescriptors⚠

cudnnGetFusedOpsConstParamPackAttribute⚠

cudnnGetFusedOpsVariantParamPackAttribute⚠

cudnnGetLRNDescriptor⚠

cudnnGetMultiHeadAttnBuffers⚠

cudnnGetMultiHeadAttnWeights⚠

cudnnGetNormalizationBackwardWorkspaceSize⚠

cudnnGetNormalizationForwardTrainingWorkspaceSize⚠

cudnnGetNormalizationTrainingReserveSpaceSize⚠

cudnnGetOpTensorDescriptor⚠

cudnnGetPooling2dDescriptor⚠

cudnnGetPooling2dForwardOutputDim⚠

cudnnGetPoolingNdDescriptor⚠

cudnnGetPoolingNdForwardOutputDim⚠

cudnnGetProperty⚠

cudnnGetRNNBackwardDataAlgorithmMaxCount⚠

cudnnGetRNNBackwardWeightsAlgorithmMaxCount⚠

cudnnGetRNNBiasMode⚠

cudnnGetRNNDataDescriptor⚠

cudnnGetRNNDescriptor_v6⚠

cudnnGetRNNDescriptor_v8⚠

cudnnGetRNNForwardInferenceAlgorithmMaxCount⚠

cudnnGetRNNForwardTrainingAlgorithmMaxCount⚠

cudnnGetRNNLinLayerBiasParams⚠

cudnnGetRNNLinLayerMatrixParams⚠

cudnnGetRNNMatrixMathType⚠

cudnnGetRNNPaddingMode⚠

cudnnGetRNNParamsSize⚠

cudnnGetRNNProjectionLayers⚠

cudnnGetRNNTempSpaceSizes⚠

cudnnGetRNNTrainingReserveSize⚠

cudnnGetRNNWeightParams⚠

cudnnGetRNNWeightSpaceSize⚠

cudnnGetRNNWorkspaceSize⚠

cudnnGetReduceTensorDescriptor⚠

cudnnGetReductionIndicesSize⚠

cudnnGetReductionWorkspaceSize⚠

cudnnGetSeqDataDescriptor⚠

cudnnGetStream⚠

cudnnGetTensor4dDescriptor⚠

cudnnGetTensorNdDescriptor⚠

cudnnGetTensorSizeInBytes⚠

cudnnGetTensorTransformDescriptor⚠

cudnnGetVersion⚠

cudnnIm2Col⚠

cudnnInitTransformDest⚠

cudnnLRNCrossChannelBackward⚠

cudnnLRNCrossChannelForward⚠

cudnnMakeFusedOpsPlan⚠

cudnnMultiHeadAttnBackwardData⚠

cudnnMultiHeadAttnBackwardWeights⚠

cudnnMultiHeadAttnForward⚠

cudnnNormalizationBackward⚠

cudnnNormalizationForwardInference⚠

cudnnNormalizationForwardTraining⚠

cudnnOpTensor⚠

cudnnOpsInferVersionCheck⚠

cudnnOpsTrainVersionCheck⚠

cudnnPoolingBackward⚠

cudnnPoolingForward⚠

cudnnQueryRuntimeError⚠

cudnnRNNBackwardData⚠

cudnnRNNBackwardDataEx⚠

cudnnRNNBackwardData_v8⚠

cudnnRNNBackwardWeights⚠

cudnnRNNBackwardWeightsEx⚠

cudnnRNNBackwardWeights_v8⚠

cudnnRNNForward⚠

cudnnRNNForwardInference⚠

cudnnRNNForwardInferenceEx⚠

cudnnRNNForwardTraining⚠

cudnnRNNForwardTrainingEx⚠

cudnnRNNGetClip⚠

cudnnRNNGetClip_v8⚠

cudnnRNNSetClip⚠

cudnnRNNSetClip_v8⚠

cudnnReduceTensor⚠

cudnnReorderFilterAndBias⚠

cudnnRestoreAlgorithm⚠

cudnnRestoreDropoutDescriptor⚠

cudnnSaveAlgorithm⚠

cudnnScaleTensor⚠

cudnnSetActivationDescriptor⚠

cudnnSetActivationDescriptorSwishBeta⚠

cudnnSetAlgorithmDescriptor⚠

cudnnSetAlgorithmPerformance⚠

cudnnSetAttnDescriptor⚠

cudnnSetCTCLossDescriptor⚠

cudnnSetCTCLossDescriptorEx⚠

cudnnSetCTCLossDescriptor_v8⚠

cudnnSetCallback⚠

cudnnSetConvolution2dDescriptor⚠

cudnnSetConvolutionGroupCount⚠

cudnnSetConvolutionMathType⚠

cudnnSetConvolutionNdDescriptor⚠

cudnnSetConvolutionReorderType⚠

cudnnSetDropoutDescriptor⚠

cudnnSetFilter4dDescriptor⚠

cudnnSetFilterNdDescriptor⚠

cudnnSetFusedOpsConstParamPackAttribute⚠

cudnnSetFusedOpsVariantParamPackAttribute⚠

cudnnSetLRNDescriptor⚠

cudnnSetOpTensorDescriptor⚠

cudnnSetPersistentRNNPlan⚠

cudnnSetPooling2dDescriptor⚠

cudnnSetPoolingNdDescriptor⚠

cudnnSetRNNAlgorithmDescriptor⚠

cudnnSetRNNBiasMode⚠

cudnnSetRNNDataDescriptor⚠

cudnnSetRNNDescriptor_v6⚠

cudnnSetRNNDescriptor_v8⚠

cudnnSetRNNMatrixMathType⚠

cudnnSetRNNPaddingMode⚠

cudnnSetRNNProjectionLayers⚠

cudnnSetReduceTensorDescriptor⚠

cudnnSetSeqDataDescriptor⚠

cudnnSetSpatialTransformerNdDescriptor⚠

cudnnSetStream⚠

cudnnSetTensor⚠

cudnnSetTensor4dDescriptor⚠

cudnnSetTensor4dDescriptorEx⚠

cudnnSetTensorNdDescriptor⚠

cudnnSetTensorNdDescriptorEx⚠

cudnnSetTensorTransformDescriptor⚠

cudnnSoftmaxBackward⚠

cudnnSoftmaxForward⚠

cudnnSpatialTfGridGeneratorBackward⚠

cudnnSpatialTfGridGeneratorForward⚠

cudnnSpatialTfSamplerBackward⚠

cudnnSpatialTfSamplerForward⚠

cudnnTransformFilter⚠

cudnnTransformTensor⚠

cudnnTransformTensorEx⚠

tensor_vec_id_c

Return C Handle for a Vector of Tensor Descriptors

Type Aliases

CUuuid

__int32_t

__int64_t

__uint32_t

cudaArray_const_t

cudaArray_t

cudaEvent_t

cudaExternalMemory_t

cudaExternalSemaphore_t

cudaFunction_t

cudaGraphExec_t

cudaGraphNode_t

cudaGraph_t

cudaGraphicsResource_t

cudaHostFn_t

cudaIpcEventHandle_t

cudaIpcMemHandle_t

cudaMemPool_t

cudaMipmappedArray_const_t

cudaMipmappedArray_t

cudaStreamCallback_t

cudaStream_t

cudaSurfaceObject_t

cudaTextureObject_t

cudaUUID_t

cudaUserObject_t

cudnnActivationDescriptor_t

cudnnAlgorithmDescriptor_t

cudnnAlgorithmPerformance_t

cudnnAlgorithm_t

cudnnAttnDescriptor_t

cudnnAttnQueryMap_t

cudnnBackendDescriptor_t

cudnnCTCLossDescriptor_t

cudnnCallback_t

cudnnConvolutionBwdDataAlgoPerf_t

cudnnConvolutionBwdFilterAlgoPerf_t

cudnnConvolutionDescriptor_t

cudnnConvolutionFwdAlgoPerf_t

cudnnDebug_t

cudnnDropoutDescriptor_t

cudnnFilterDescriptor_t

cudnnFusedOpsConstParamPack_t

cudnnFusedOpsPlan_t

cudnnFusedOpsVariantParamPack_t

cudnnHandle_t

cudnnLRNDescriptor_t

cudnnOpTensorDescriptor_t

cudnnPersistentRNNPlan_t

cudnnPoolingDescriptor_t

cudnnRNNDataDescriptor_t

cudnnRNNDescriptor_t

cudnnRNNPaddingMode_t

cudnnReduceTensorDescriptor_t

cudnnSeqDataDescriptor_t

cudnnSpatialTransformerDescriptor_t

cudnnTensorDescriptor_t

cudnnTensorTransformDescriptor_t

libraryPropertyType

libraryPropertyType_t

Unions

cudaExternalMemoryHandleDesc__bindgen_ty_1

cudaExternalSemaphoreHandleDesc__bindgen_ty_1

cudaExternalSemaphoreSignalParams__bindgen_ty_1__bindgen_ty_2

cudaExternalSemaphoreSignalParams_v1__bindgen_ty_1__bindgen_ty_2

cudaExternalSemaphoreWaitParams__bindgen_ty_1__bindgen_ty_2

cudaExternalSemaphoreWaitParams_v1__bindgen_ty_1__bindgen_ty_2

cudaKernelNodeAttrValue

cudaResourceDesc__bindgen_ty_1

cudaStreamAttrValue

cudnnAlgorithmUnionStruct_Algorithm