Function rcudnn::cudaDeviceGetLimit

pub unsafe extern "C" fn cudaDeviceGetLimit(
    pValue: *mut usize, 
    limit: cudaLimit
) -> cudaError

\brief Returns resource limits

Returns in \p *pValue the current size of \p limit. The supported ::cudaLimit values are:

::cudaLimitStackSize: stack size in bytes of each GPU thread;

::cudaLimitPrintfFifoSize: size in bytes of the shared FIFO used by the ::printf() device system call.

::cudaLimitMallocHeapSize: size in bytes of the heap used by the ::malloc() and ::free() device system calls;

::cudaLimitDevRuntimeSyncDepth: maximum grid depth at which a thread can isssue the device runtime call ::cudaDeviceSynchronize() to wait on child grid launches to complete.

::cudaLimitDevRuntimePendingLaunchCount: maximum number of outstanding device runtime launches.

::cudaLimitMaxL2FetchGranularity: L2 cache fetch granularity.

::cudaLimitPersistingL2CacheSize: Persisting L2 cache size in bytes

\param limit - Limit to query \param pValue - Returned size of the limit

\return ::cudaSuccess, ::cudaErrorUnsupportedLimit, ::cudaErrorInvalidValue \notefnerr \note_init_rt \note_callback

\sa ::cudaDeviceSetLimit, ::cuCtxGetLimit