Function rcudnn::cudaDeviceGetStreamPriorityRange

pub unsafe extern "C" fn cudaDeviceGetStreamPriorityRange(
    leastPriority: *mut i32, 
    greatestPriority: *mut i32
) -> cudaError

\brief Returns numerical values that correspond to the least and greatest stream priorities.

Returns in \p *leastPriority and \p *greatestPriority the numerical values that correspond to the least and greatest stream priorities respectively. Stream priorities follow a convention where lower numbers imply greater priorities. The range of meaningful stream priorities is given by [\p *greatestPriority, \p *leastPriority]. If the user attempts to create a stream with a priority value that is outside the the meaningful range as specified by this API, the priority is automatically clamped down or up to either \p *leastPriority or \p *greatestPriority respectively. See ::cudaStreamCreateWithPriority for details on creating a priority stream. A NULL may be passed in for \p *leastPriority or \p *greatestPriority if the value is not desired.

This function will return ‘0’ in both \p *leastPriority and \p *greatestPriority if the current context’s device does not support stream priorities (see ::cudaDeviceGetAttribute).

\param leastPriority - Pointer to an int in which the numerical value for least stream priority is returned \param greatestPriority - Pointer to an int in which the numerical value for greatest stream priority is returned

\return ::cudaSuccess \notefnerr \note_init_rt \note_callback

\sa ::cudaStreamCreateWithPriority, ::cudaStreamGetPriority, ::cuCtxGetStreamPriorityRange