Function rcudnn::cudaEventCreateWithFlags

pub unsafe extern "C" fn cudaEventCreateWithFlags(
    event: *mut *mut CUevent_st, 
    flags: u32
) -> cudaError

\brief Creates an event object with the specified flags

Creates an event object for the current device with the specified flags. Valid flags include:

::cudaEventDefault: Default event creation flag.

::cudaEventBlockingSync: Specifies that event should use blocking synchronization. A host thread that uses ::cudaEventSynchronize() to wait on an event created with this flag will block until the event actually completes.

::cudaEventDisableTiming: Specifies that the created event does not need to record timing data. Events created with this flag specified and the ::cudaEventBlockingSync flag not specified will provide the best performance when used with ::cudaStreamWaitEvent() and ::cudaEventQuery().

::cudaEventInterprocess: Specifies that the created event may be used as an interprocess event by ::cudaIpcGetEventHandle(). ::cudaEventInterprocess must be specified along with ::cudaEventDisableTiming.

\param event - Newly created event \param flags - Flags for new event

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorLaunchFailure, ::cudaErrorMemoryAllocation \notefnerr \note_init_rt \note_callback

\sa \ref ::cudaEventCreate(cudaEvent_t*) “cudaEventCreate (C API)”, ::cudaEventSynchronize, ::cudaEventDestroy, ::cudaEventElapsedTime, ::cudaStreamWaitEvent, ::cuEventCreate