Function rcudnn::cudaGetDriverEntryPoint

pub unsafe extern "C" fn cudaGetDriverEntryPoint(
    symbol: *const i8, 
    funcPtr: *mut *mut c_void, 
    flags: u64
) -> cudaError

\brief Returns the requested driver API function pointer

Returns in \p **funcPtr the address of the CUDA driver function for the requested flags.

For a requested driver symbol, if the CUDA version in which the driver symbol was introduced is less than or equal to the CUDA runtime version, the API will return the function pointer to the corresponding versioned driver function.

The pointer returned by the API should be cast to a function pointer matching the requested driver function’s definition in the API header file. The function pointer typedef can be picked up from the corresponding typedefs header file. For example, cudaTypedefs.h consists of function pointer typedefs for driver APIs defined in cuda.h.

The API will return ::cudaErrorSymbolNotFound if the requested driver function is not supported on the platform, no ABI compatible driver function exists for the CUDA runtime version or if the driver symbol is invalid.

The requested flags can be:

::cudaEnableDefault: This is the default mode. This is equivalent to ::cudaEnablePerThreadDefaultStream if the code is compiled with –default-stream per-thread compilation flag or the macro CUDA_API_PER_THREAD_DEFAULT_STREAM is defined; ::cudaEnableLegacyStream otherwise.

::cudaEnableLegacyStream: This will enable the search for all driver symbols that match the requested driver symbol name except the corresponding per-thread versions.

::cudaEnablePerThreadDefaultStream: This will enable the search for all driver symbols that match the requested driver symbol name including the per-thread versions. If a per-thread version is not found, the API will return the legacy version of the driver function.

\param symbol - The base name of the driver API function to look for. As an example, for the driver API ::cuMemAlloc_v2, \p symbol would be cuMemAlloc. Note that the API will use the CUDA runtime version to return the address to the most recent ABI compatible driver symbol, ::cuMemAlloc or ::cuMemAlloc_v2. \param funcPtr - Location to return the function pointer to the requested driver function \param flags - Flags to specify search options.

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorNotSupported, ::cudaErrorSymbolNotFound \note_version_mixing \note_init_rt \note_callback

\sa ::cuGetProcAddress