Function rcudnn::cudaGetSymbolSize

pub unsafe extern "C" fn cudaGetSymbolSize(
    size: *mut usize, 
    symbol: *const c_void
) -> cudaError

\brief Finds the size of the object associated with a CUDA symbol

Returns in \p *size the size of symbol \p symbol. \p symbol is a variable that resides in global or constant memory space. If \p symbol cannot be found, or if \p symbol is not declared in global or constant memory space, \p *size is unchanged and the error ::cudaErrorInvalidSymbol is returned.

\param size - Size of object associated with symbol \param symbol - Device symbol address

\return ::cudaSuccess, ::cudaErrorInvalidSymbol, ::cudaErrorNoKernelImageForDevice \notefnerr \note_string_api_deprecation \note_init_rt \note_callback

\sa \ref ::cudaGetSymbolAddress(void**, const void*) “cudaGetSymbolAddress (C API)”, \ref ::cudaGetSymbolSize(size_t*, const T&) “cudaGetSymbolSize (C++ API)”, ::cuModuleGetGlobal