Function rcudnn::cudaGraphAddEmptyNode

pub unsafe extern "C" fn cudaGraphAddEmptyNode(
    pGraphNode: *mut *mut CUgraphNode_st, 
    graph: *mut CUgraph_st, 
    pDependencies: *const *mut CUgraphNode_st, 
    numDependencies: usize
) -> cudaError

\brief Creates an empty node and adds it to a graph

Creates a new node which performs no operation, and adds it to \p graph with \p numDependencies dependencies specified via \p pDependencies. It is possible for \p numDependencies to be 0, in which case the node will be placed at the root of the graph. \p pDependencies may not have any duplicate entries. A handle to the new node will be returned in \p pGraphNode.

An empty node performs no operation during execution, but can be used for transitive ordering. For example, a phased execution graph with 2 groups of n nodes with a barrier between them can be represented using an empty node and 2*n dependency edges, rather than no empty node and n^2 dependency edges.

\param pGraphNode - Returns newly created node \param graph - Graph to which to add the node \param pDependencies - Dependencies of the node \param numDependencies - Number of dependencies

\return ::cudaSuccess, ::cudaErrorInvalidValue \note_graph_thread_safety \note_init_rt \note_callback

\sa ::cudaGraphCreate, ::cudaGraphDestroyNode, ::cudaGraphAddChildGraphNode, ::cudaGraphAddKernelNode, ::cudaGraphAddHostNode, ::cudaGraphAddMemcpyNode, ::cudaGraphAddMemsetNode