Function rcudnn::cudaGraphAddHostNode

pub unsafe extern "C" fn cudaGraphAddHostNode(
    pGraphNode: *mut *mut CUgraphNode_st, 
    graph: *mut CUgraph_st, 
    pDependencies: *const *mut CUgraphNode_st, 
    numDependencies: usize, 
    pNodeParams: *const cudaHostNodeParams
) -> cudaError

\brief Creates a host execution node and adds it to a graph

Creates a new CPU execution node and adds it to \p graph with \p numDependencies dependencies specified via \p pDependencies and arguments specified in \p pNodeParams. It is possible for \p numDependencies to be 0, in which case the node will be placed at the root of the graph. \p pDependencies may not have any duplicate entries. A handle to the new node will be returned in \p pGraphNode.

When the graph is launched, the node will invoke the specified CPU function. Host nodes are not supported under MPS with pre-Volta GPUs.

\param pGraphNode - Returns newly created node \param graph - Graph to which to add the node \param pDependencies - Dependencies of the node \param numDependencies - Number of dependencies \param pNodeParams - Parameters for the host node

\return ::cudaSuccess, ::cudaErrorNotSupported, ::cudaErrorInvalidValue \note_graph_thread_safety \notefnerr \note_init_rt \note_callback

\sa ::cudaLaunchHostFunc, ::cudaGraphHostNodeGetParams, ::cudaGraphHostNodeSetParams, ::cudaGraphCreate, ::cudaGraphDestroyNode, ::cudaGraphAddChildGraphNode, ::cudaGraphAddEmptyNode, ::cudaGraphAddKernelNode, ::cudaGraphAddMemcpyNode, ::cudaGraphAddMemsetNode