Function rcudnn::cudaGraphAddMemsetNode

pub unsafe extern "C" fn cudaGraphAddMemsetNode(
    pGraphNode: *mut *mut CUgraphNode_st, 
    graph: *mut CUgraph_st, 
    pDependencies: *const *mut CUgraphNode_st, 
    numDependencies: usize, 
    pMemsetParams: *const cudaMemsetParams
) -> cudaError

\brief Creates a memset node and adds it to a graph

Creates a new memset node and adds it to \p graph with \p numDependencies dependencies specified via \p pDependencies. It is possible for \p numDependencies to be 0, in which case the node will be placed at the root of the graph. \p pDependencies may not have any duplicate entries. A handle to the new node will be returned in \p pGraphNode.

The element size must be 1, 2, or 4 bytes. When the graph is launched, the node will perform the memset described by \p pMemsetParams.

\param pGraphNode - Returns newly created node \param graph - Graph to which to add the node \param pDependencies - Dependencies of the node \param numDependencies - Number of dependencies \param pMemsetParams - Parameters for the memory set

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidDevice \note_graph_thread_safety \notefnerr \note_init_rt \note_callback

\sa ::cudaMemset2D, ::cudaGraphMemsetNodeGetParams, ::cudaGraphMemsetNodeSetParams, ::cudaGraphCreate, ::cudaGraphDestroyNode, ::cudaGraphAddChildGraphNode, ::cudaGraphAddEmptyNode, ::cudaGraphAddKernelNode, ::cudaGraphAddHostNode, ::cudaGraphAddMemcpyNode