Function rcudnn::cudaGraphKernelNodeGetParams

pub unsafe extern "C" fn cudaGraphKernelNodeGetParams(
    node: *mut CUgraphNode_st, 
    pNodeParams: *mut cudaKernelNodeParams
) -> cudaError

\brief Returns a kernel node’s parameters

Returns the parameters of kernel node \p node in \p pNodeParams. The \p kernelParams or \p extra array returned in \p pNodeParams, as well as the argument values it points to, are owned by the node. This memory remains valid until the node is destroyed or its parameters are modified, and should not be modified directly. Use ::cudaGraphKernelNodeSetParams to update the parameters of this node.

The params will contain either \p kernelParams or \p extra, according to which of these was most recently set on the node.

\param node - Node to get the parameters for \param pNodeParams - Pointer to return the parameters

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidDeviceFunction \note_graph_thread_safety \notefnerr \note_init_rt \note_callback

\sa ::cudaLaunchKernel, ::cudaGraphAddKernelNode, ::cudaGraphKernelNodeSetParams