Function rcudnn::cudaMemPoolCreate

pub unsafe extern "C" fn cudaMemPoolCreate(
    memPool: *mut *mut CUmemPoolHandle_st, 
    poolProps: *const cudaMemPoolProps
) -> cudaError

\brief Creates a memory pool

Creates a CUDA memory pool and returns the handle in \p pool. The \p poolProps determines the properties of the pool such as the backing device and IPC capabilities.

By default, the pool’s memory will be accessible from the device it is allocated on.

\note Specifying cudaMemHandleTypeNone creates a memory pool that will not support IPC.

\returns ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorNotSupported

\sa ::cuMemPoolCreate, ::cudaDeviceSetMemPool, ::cudaMallocFromPoolAsync, ::cudaMemPoolExportToShareableHandle, ::cudaDeviceGetDefaultMemPool, ::cudaDeviceGetMemPool