Function rcudnn::cudaMemcpy2DArrayToArray

pub unsafe extern "C" fn cudaMemcpy2DArrayToArray(
    dst: *mut cudaArray, 
    wOffsetDst: usize, 
    hOffsetDst: usize, 
    src: *const cudaArray, 
    wOffsetSrc: usize, 
    hOffsetSrc: usize, 
    width: usize, 
    height: usize, 
    kind: cudaMemcpyKind
) -> cudaError

\brief Copies data between host and device

Copies a matrix (\p height rows of \p width bytes each) from the CUDA array \p src starting at \p hOffsetSrc rows and \p wOffsetSrc bytes from the upper left corner to the CUDA array \p dst starting at \p hOffsetDst rows and \p wOffsetDst bytes from the upper left corner, where \p kind specifies the direction of the copy, and must be one of ::cudaMemcpyHostToHost, ::cudaMemcpyHostToDevice, ::cudaMemcpyDeviceToHost, ::cudaMemcpyDeviceToDevice, or ::cudaMemcpyDefault. Passing ::cudaMemcpyDefault is recommended, in which case the type of transfer is inferred from the pointer values. However, ::cudaMemcpyDefault is only allowed on systems that support unified virtual addressing. \p wOffsetDst + \p width must not exceed the width of the CUDA array \p dst. \p wOffsetSrc + \p width must not exceed the width of the CUDA array \p src.

\param dst - Destination memory address \param wOffsetDst - Destination starting X offset (columns in bytes) \param hOffsetDst - Destination starting Y offset (rows) \param src - Source memory address \param wOffsetSrc - Source starting X offset (columns in bytes) \param hOffsetSrc - Source starting Y offset (rows) \param width - Width of matrix transfer (columns in bytes) \param height - Height of matrix transfer (rows) \param kind - Type of transfer

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidMemcpyDirection \notefnerr \note_sync \note_init_rt \note_callback

\sa ::cudaMemcpy, ::cudaMemcpy2D, ::cudaMemcpy2DToArray, ::cudaMemcpy2DFromArray, ::cudaMemcpyToSymbol, ::cudaMemcpyFromSymbol, ::cudaMemcpyAsync, ::cudaMemcpy2DAsync, ::cudaMemcpy2DToArrayAsync, ::cudaMemcpy2DFromArrayAsync, ::cudaMemcpyToSymbolAsync, ::cudaMemcpyFromSymbolAsync, ::cuMemcpy2D, ::cuMemcpy2DUnaligned