Function rcudnn::cudaMemcpy2DFromArrayAsync

pub unsafe extern "C" fn cudaMemcpy2DFromArrayAsync(
    dst: *mut c_void, 
    dpitch: usize, 
    src: *const cudaArray, 
    wOffset: usize, 
    hOffset: usize, 
    width: usize, 
    height: usize, 
    kind: cudaMemcpyKind, 
    stream: *mut CUstream_st
) -> cudaError

\brief Copies data between host and device

Copies a matrix (\p height rows of \p width bytes each) from the CUDA array \p srcArray starting at the upper left corner (\p wOffset, \p hOffset) to the memory area pointed to by \p dst, where \p kind specifies the direction of the copy, and must be one of ::cudaMemcpyHostToHost, ::cudaMemcpyHostToDevice, ::cudaMemcpyDeviceToHost, ::cudaMemcpyDeviceToDevice, or ::cudaMemcpyDefault. Passing ::cudaMemcpyDefault is recommended, in which case the type of transfer is inferred from the pointer values. However, ::cudaMemcpyDefault is only allowed on systems that support unified virtual addressing. \p dpitch is the width in memory in bytes of the 2D array pointed to by \p dst, including any padding added to the end of each row. \p wOffset + \p width must not exceed the width of the CUDA array \p src. \p width must not exceed \p dpitch. ::cudaMemcpy2DFromArrayAsync() returns an error if \p dpitch exceeds the maximum allowed.

::cudaMemcpy2DFromArrayAsync() is asynchronous with respect to the host, so the call may return before the copy is complete. The copy can optionally be associated to a stream by passing a non-zero \p stream argument. If \p kind is ::cudaMemcpyHostToDevice or ::cudaMemcpyDeviceToHost and \p stream is non-zero, the copy may overlap with operations in other streams.

\param dst - Destination memory address \param dpitch - Pitch of destination memory \param src - Source memory address \param wOffset - Source starting X offset \param hOffset - Source starting Y offset \param width - Width of matrix transfer (columns in bytes) \param height - Height of matrix transfer (rows) \param kind - Type of transfer \param stream - Stream identifier

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidPitchValue, ::cudaErrorInvalidMemcpyDirection \notefnerr \note_async \note_null_stream \note_init_rt \note_callback

\sa ::cudaMemcpy, ::cudaMemcpy2D, ::cudaMemcpy2DToArray, ::cudaMemcpy2DFromArray, ::cudaMemcpy2DArrayToArray, ::cudaMemcpyToSymbol, ::cudaMemcpyFromSymbol, ::cudaMemcpyAsync, ::cudaMemcpy2DAsync, ::cudaMemcpy2DToArrayAsync,

::cudaMemcpyToSymbolAsync, ::cudaMemcpyFromSymbolAsync, ::cuMemcpy2DAsync