Function rcudnn::cudaMemcpy3DPeer

pub unsafe extern "C" fn cudaMemcpy3DPeer(
    p: *const cudaMemcpy3DPeerParms
) -> cudaError

\brief Copies memory between devices

Perform a 3D memory copy according to the parameters specified in \p p. See the definition of the ::cudaMemcpy3DPeerParms structure for documentation of its parameters.

Note that this function is synchronous with respect to the host only if the source or destination of the transfer is host memory. Note also that this copy is serialized with respect to all pending and future asynchronous work in to the current device, the copy's source device, and the copy's destination device (use ::cudaMemcpy3DPeerAsync to avoid this synchronization).

\param p - Parameters for the memory copy

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidDevice \notefnerr \note_sync \note_init_rt \note_callback

\sa ::cudaMemcpy, ::cudaMemcpyPeer, ::cudaMemcpyAsync, ::cudaMemcpyPeerAsync, ::cudaMemcpy3DPeerAsync, ::cuMemcpy3DPeer