Function rcudnn::cudaMemcpyFromSymbol

pub unsafe extern "C" fn cudaMemcpyFromSymbol(
    dst: *mut c_void, 
    symbol: *const c_void, 
    count: usize, 
    offset: usize, 
    kind: cudaMemcpyKind
) -> cudaError

\brief Copies data from the given symbol on the device

Copies \p count bytes from the memory area pointed to by \p offset bytes from the start of symbol \p symbol to the memory area pointed to by \p dst. The memory areas may not overlap. \p symbol is a variable that resides in global or constant memory space. \p kind can be either ::cudaMemcpyDeviceToHost, ::cudaMemcpyDeviceToDevice, or ::cudaMemcpyDefault. Passing ::cudaMemcpyDefault is recommended, in which case the type of transfer is inferred from the pointer values. However, ::cudaMemcpyDefault is only allowed on systems that support unified virtual addressing.

\param dst - Destination memory address \param symbol - Device symbol address \param count - Size in bytes to copy \param offset - Offset from start of symbol in bytes \param kind - Type of transfer

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidSymbol, ::cudaErrorInvalidMemcpyDirection, ::cudaErrorNoKernelImageForDevice \notefnerr \note_sync \note_string_api_deprecation \note_init_rt \note_callback

\sa ::cudaMemcpy, ::cudaMemcpy2D, ::cudaMemcpy2DToArray, ::cudaMemcpy2DFromArray, ::cudaMemcpy2DArrayToArray, ::cudaMemcpyToSymbol, ::cudaMemcpyAsync, ::cudaMemcpy2DAsync, ::cudaMemcpy2DToArrayAsync, ::cudaMemcpy2DFromArrayAsync, ::cudaMemcpyToSymbolAsync, ::cudaMemcpyFromSymbolAsync, ::cuMemcpy, ::cuMemcpyDtoH, ::cuMemcpyDtoD