Function rcudnn::cudaMemcpyFromSymbolAsync

pub unsafe extern "C" fn cudaMemcpyFromSymbolAsync(
    dst: *mut c_void, 
    symbol: *const c_void, 
    count: usize, 
    offset: usize, 
    kind: cudaMemcpyKind, 
    stream: *mut CUstream_st
) -> cudaError

\brief Copies data from the given symbol on the device

Copies \p count bytes from the memory area pointed to by \p offset bytes from the start of symbol \p symbol to the memory area pointed to by \p dst. The memory areas may not overlap. \p symbol is a variable that resides in global or constant memory space. \p kind can be either ::cudaMemcpyDeviceToHost, ::cudaMemcpyDeviceToDevice, or ::cudaMemcpyDefault. Passing ::cudaMemcpyDefault is recommended, in which case the type of transfer is inferred from the pointer values. However, ::cudaMemcpyDefault is only allowed on systems that support unified virtual addressing.

::cudaMemcpyFromSymbolAsync() is asynchronous with respect to the host, so the call may return before the copy is complete. The copy can optionally be associated to a stream by passing a non-zero \p stream argument. If \p kind is ::cudaMemcpyDeviceToHost and \p stream is non-zero, the copy may overlap with operations in other streams.

\param dst - Destination memory address \param symbol - Device symbol address \param count - Size in bytes to copy \param offset - Offset from start of symbol in bytes \param kind - Type of transfer \param stream - Stream identifier

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidSymbol, ::cudaErrorInvalidMemcpyDirection, ::cudaErrorNoKernelImageForDevice \notefnerr \note_async \note_null_stream \note_string_api_deprecation \note_init_rt \note_callback

\sa ::cudaMemcpy, ::cudaMemcpy2D, ::cudaMemcpy2DToArray, ::cudaMemcpy2DFromArray, ::cudaMemcpy2DArrayToArray, ::cudaMemcpyToSymbol, ::cudaMemcpyFromSymbol, ::cudaMemcpyAsync, ::cudaMemcpy2DAsync, ::cudaMemcpy2DToArrayAsync, ::cudaMemcpy2DFromArrayAsync, ::cudaMemcpyToSymbolAsync, ::cuMemcpyAsync, ::cuMemcpyDtoHAsync, ::cuMemcpyDtoDAsync