Function rcudnn::cudaMemset

pub unsafe extern "C" fn cudaMemset(
    devPtr: *mut c_void, 
    value: i32, 
    count: usize
) -> cudaError

\brief Initializes or sets device memory to a value

Fills the first \p count bytes of the memory area pointed to by \p devPtr with the constant byte value \p value.

Note that this function is asynchronous with respect to the host unless \p devPtr refers to pinned host memory.

\param devPtr - Pointer to device memory \param value - Value to set for each byte of specified memory \param count - Size in bytes to set

\return ::cudaSuccess, ::cudaErrorInvalidValue, \notefnerr \note_memset \note_init_rt \note_callback

\sa ::cuMemsetD8, ::cuMemsetD16, ::cuMemsetD32