Function rcudnn::cudaStreamCreateWithFlags

pub unsafe extern "C" fn cudaStreamCreateWithFlags(
    pStream: *mut *mut CUstream_st, 
    flags: u32
) -> cudaError

\brief Create an asynchronous stream

Creates a new asynchronous stream. The \p flags argument determines the behaviors of the stream. Valid values for \p flags are

::cudaStreamDefault: Default stream creation flag.

::cudaStreamNonBlocking: Specifies that work running in the created stream may run concurrently with work in stream 0 (the NULL stream), and that the created stream should perform no implicit synchronization with stream 0.

\param pStream - Pointer to new stream identifier \param flags - Parameters for stream creation

\return ::cudaSuccess, ::cudaErrorInvalidValue \notefnerr \note_init_rt \note_callback

\sa ::cudaStreamCreate, ::cudaStreamCreateWithPriority, ::cudaStreamGetFlags, ::cudaStreamQuery, ::cudaStreamSynchronize, ::cudaStreamWaitEvent, ::cudaStreamAddCallback, ::cudaStreamDestroy, ::cuStreamCreate