Function rcudnn::cudaStreamDestroy

pub unsafe extern "C" fn cudaStreamDestroy(
    stream: *mut CUstream_st
) -> cudaError

\brief Destroys and cleans up an asynchronous stream

Destroys and cleans up the asynchronous stream specified by \p stream.

In case the device is still doing work in the stream \p stream when ::cudaStreamDestroy() is called, the function will return immediately and the resources associated with \p stream will be released automatically once the device has completed all work in \p stream.

\param stream - Stream identifier

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorInvalidResourceHandle \note_null_stream \notefnerr \note_init_rt \note_callback \note_destroy_ub

\sa ::cudaStreamCreate, ::cudaStreamCreateWithFlags, ::cudaStreamQuery, ::cudaStreamWaitEvent, ::cudaStreamSynchronize, ::cudaStreamAddCallback, ::cuStreamDestroy