Function rcudnn::cudaStreamEndCapture

pub unsafe extern "C" fn cudaStreamEndCapture(
    stream: *mut CUstream_st, 
    pGraph: *mut *mut CUgraph_st
) -> cudaError

\brief Ends capture on a stream, returning the captured graph

End capture on \p stream, returning the captured graph via \p pGraph. Capture must have been initiated on \p stream via a call to ::cudaStreamBeginCapture. If capture was invalidated, due to a violation of the rules of stream capture, then a NULL graph will be returned.

If the \p mode argument to ::cudaStreamBeginCapture was not ::cudaStreamCaptureModeRelaxed, this call must be from the same thread as ::cudaStreamBeginCapture.

\param stream - Stream to query \param pGraph - The captured graph

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorStreamCaptureWrongThread \notefnerr

\sa ::cudaStreamCreate, ::cudaStreamBeginCapture, ::cudaStreamIsCapturing