Function rcudnn::cudaStreamUpdateCaptureDependencies

pub unsafe extern "C" fn cudaStreamUpdateCaptureDependencies(
    stream: *mut CUstream_st, 
    dependencies: *mut *mut CUgraphNode_st, 
    numDependencies: usize, 
    flags: u32
) -> cudaError

\brief Update the set of dependencies in a capturing stream (11.3+)

Modifies the dependency set of a capturing stream. The dependency set is the set of nodes that the next captured node in the stream will depend on.

Valid flags are ::cudaStreamAddCaptureDependencies and ::cudaStreamSetCaptureDependencies. These control whether the set passed to the API is added to the existing set or replaces it. A flags value of 0 defaults to ::cudaStreamAddCaptureDependencies.

Nodes that are removed from the dependency set via this API do not result in ::cudaErrorStreamCaptureUnjoined if they are unreachable from the stream at ::cudaStreamEndCapture.

Returns ::cudaErrorIllegalState if the stream is not capturing.

This API is new in CUDA 11.3. Developers requiring compatibility across minor versions of the CUDA driver to 11.0 should not use this API or provide a fallback.

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorIllegalState \notefnerr

\sa ::cudaStreamBeginCapture, ::cudaStreamGetCaptureInfo, ::cudaStreamGetCaptureInfo_v2