Function rcudnn::cudaThreadSynchronize

pub unsafe extern "C" fn cudaThreadSynchronize() -> cudaError

\brief Wait for compute device to finish

\deprecated

Note that this function is deprecated because its name does not reflect its behavior. Its functionality is similar to the non-deprecated function ::cudaDeviceSynchronize(), which should be used instead.

Blocks until the device has completed all preceding requested tasks. ::cudaThreadSynchronize() returns an error if one of the preceding tasks has failed. If the ::cudaDeviceScheduleBlockingSync flag was set for this device, the host thread will block until the device has finished its work.

\return ::cudaSuccess \notefnerr \note_init_rt \note_callback

\sa ::cudaDeviceSynchronize