Struct rcudnn::cudaFuncAttributes

#[repr(C)]
pub struct cudaFuncAttributes {
    pub sharedSizeBytes: usize,
    pub constSizeBytes: usize,
    pub localSizeBytes: usize,
    pub maxThreadsPerBlock: i32,
    pub numRegs: i32,
    pub ptxVersion: i32,
    pub binaryVersion: i32,
    pub cacheModeCA: i32,
    pub maxDynamicSharedSizeBytes: i32,
    pub preferredShmemCarveout: i32,
}

CUDA function attributes

Fields
sharedSizeBytes: usize

The size in bytes of statically-allocated shared memory per block required by this function. This does not include dynamically-allocated shared memory requested by the user at runtime.

constSizeBytes: usize

The size in bytes of user-allocated constant memory required by this function.

localSizeBytes: usize

The size in bytes of local memory used by each thread of this function.

maxThreadsPerBlock: i32

The maximum number of threads per block, beyond which a launch of the function would fail. This number depends on both the function and the device on which the function is currently loaded.

numRegs: i32

The number of registers used by each thread of this function.

ptxVersion: i32

The PTX virtual architecture version for which the function was compiled. This value is the major PTX version * 10 + the minor PTX version, so a PTX version 1.3 function would return the value 13.

binaryVersion: i32

The binary architecture version for which the function was compiled. This value is the major binary version * 10 + the minor binary version, so a binary version 1.3 function would return the value 13.

cacheModeCA: i32

The attribute to indicate whether the function has been compiled with user specified option “-Xptxas –dlcm=ca” set.

maxDynamicSharedSizeBytes: i32

The maximum size in bytes of dynamic shared memory per block for this function. Any launch must have a dynamic shared memory size smaller than this value.

preferredShmemCarveout: i32

On devices where the L1 cache and shared memory use the same hardware resources, this sets the shared memory carveout preference, in percent of the maximum shared memory. Refer to ::cudaDevAttrMaxSharedMemoryPerMultiprocessor. This is only a hint, and the driver can choose a different ratio if required to execute the function. See ::cudaFuncSetAttribute