Function rcudnn::cudaGraphInstantiate

pub unsafe extern "C" fn cudaGraphInstantiate(
    pGraphExec: *mut *mut CUgraphExec_st, 
    graph: *mut CUgraph_st, 
    pErrorNode: *mut *mut CUgraphNode_st, 
    pLogBuffer: *mut i8, 
    bufferSize: usize
) -> cudaError

\brief Creates an executable graph from a graph

Instantiates \p graph as an executable graph. The graph is validated for any structural constraints or intra-node constraints which were not previously validated. If instantiation is successful, a handle to the instantiated graph is returned in \p pGraphExec.

If there are any errors, diagnostic information may be returned in \p pErrorNode and \p pLogBuffer. This is the primary way to inspect instantiation errors. The output will be null terminated unless the diagnostics overflow the buffer. In this case, they will be truncated, and the last byte can be inspected to determine if truncation occurred.

\param pGraphExec - Returns instantiated graph \param graph - Graph to instantiate \param pErrorNode - In case of an instantiation error, this may be modified to indicate a node contributing to the error \param pLogBuffer - A character buffer to store diagnostic messages \param bufferSize - Size of the log buffer in bytes

\return ::cudaSuccess, ::cudaErrorInvalidValue \note_graph_thread_safety \notefnerr \note_init_rt \note_callback

\sa ::cudaGraphCreate, ::cudaGraphLaunch, ::cudaGraphExecDestroy