Function rcudnn::cudaLaunchKernel

pub unsafe extern "C" fn cudaLaunchKernel(
    func: *const c_void, 
    gridDim: dim3, 
    blockDim: dim3, 
    args: *mut *mut c_void, 
    sharedMem: usize, 
    stream: *mut CUstream_st
) -> cudaError

\brief Launches a device function

The function invokes kernel \p func on \p gridDim (\p gridDim.x × \p gridDim.y × \p gridDim.z) grid of blocks. Each block contains \p blockDim (\p blockDim.x × \p blockDim.y × \p blockDim.z) threads.

If the kernel has N parameters the \p args should point to array of N pointers. Each pointer, from args[0] to args[N - 1], point to the region of memory from which the actual parameter will be copied.

For templated functions, pass the function symbol as follows: func_name<template_arg_0,…,template_arg_N>

\p sharedMem sets the amount of dynamic shared memory that will be available to each thread block.

\p stream specifies a stream the invocation is associated to.

\param func - Device function symbol \param gridDim - Grid dimentions \param blockDim - Block dimentions \param args - Arguments \param sharedMem - Shared memory \param stream - Stream identifier

\return ::cudaSuccess, ::cudaErrorInvalidDeviceFunction, ::cudaErrorInvalidConfiguration, ::cudaErrorLaunchFailure, ::cudaErrorLaunchTimeout, ::cudaErrorLaunchOutOfResources, ::cudaErrorSharedObjectInitFailed, ::cudaErrorInvalidPtx, ::cudaErrorUnsupportedPtxVersion, ::cudaErrorNoKernelImageForDevice, ::cudaErrorJitCompilerNotFound, ::cudaErrorJitCompilationDisabled \note_null_stream \notefnerr \note_init_rt \note_callback

\sa \ref ::cudaLaunchKernel(const T *func, dim3 gridDim, dim3 blockDim, void **args, size_t sharedMem, cudaStream_t stream) “cudaLaunchKernel (C++ API)”, ::cuLaunchKernel