Function rcudnn::cudaStreamIsCapturing

pub unsafe extern "C" fn cudaStreamIsCapturing(
    stream: *mut CUstream_st, 
    pCaptureStatus: *mut cudaStreamCaptureStatus
) -> cudaError

\brief Returns a stream’s capture status

Return the capture status of \p stream via \p pCaptureStatus. After a successful call, \p *pCaptureStatus will contain one of the following:

::cudaStreamCaptureStatusNone: The stream is not capturing.

::cudaStreamCaptureStatusActive: The stream is capturing.

::cudaStreamCaptureStatusInvalidated: The stream was capturing but an error has invalidated the capture sequence. The capture sequence must be terminated with ::cudaStreamEndCapture on the stream where it was initiated in order to continue using \p stream.

Note that, if this is called on ::cudaStreamLegacy (the “null stream”) while a blocking stream on the same device is capturing, it will return ::cudaErrorStreamCaptureImplicit and \p *pCaptureStatus is unspecified after the call. The blocking stream capture is not invalidated.

When a blocking stream is capturing, the legacy stream is in an unusable state until the blocking stream capture is terminated. The legacy stream is not supported for stream capture, but attempted use would have an implicit dependency on the capturing stream(s).

\param stream - Stream to query \param pCaptureStatus - Returns the stream’s capture status

\return ::cudaSuccess, ::cudaErrorInvalidValue, ::cudaErrorStreamCaptureImplicit \notefnerr

\sa ::cudaStreamCreate, ::cudaStreamBeginCapture, ::cudaStreamEndCapture